Добавление тестов для LCG

+++ src/main/kotlin/algs/LCG.kt
@@ -7,14 +7,14 @@ class LCG(
     fun getState(): Int { return state_in }
 
     fun LCG_next(): Int {
-        val a = coeffs_in[0]
-        val c = coeffs_in[1]
-        val m = coeffs_in[2]
+        val a = coeffs_in[0].toLong()
+        val c = coeffs_in[1].toLong()
+        val m = coeffs_in[2].toLong()
 
         val out = (a * state_in + c) % m
 
-        state_in = out
+        state_in = out.toInt()
 
-        return out
+        return out.toInt()
     }
 }
